Refactor: Rename components to kebab-case with client/server suffixes
