+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import CryptoCardsHoverEffect from "@/components/crypto/card/CryptoCardsHoverEffect";
+import CryptoCardsHoverEffect from "@/components/crypto/card/crypto-cards-hover-effect.client";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
