+++ app/page.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from "react";
-import AssetsGrid from "@/components/crypto/grid/AssetsGrid";
+import AssetsGrid from "@/components/crypto/grid/assets-grid";
 export const dynamic = "force-dynamic";
 
 export default function Home() {
