+++ components/crypto/crypto-icon.tsx
