+++ components/crypto/grid/loading-grid.tsx
@@ -1,6 +1,6 @@
 import { GridItem } from "@/components/ui/grid";
 import { Grid } from "@/components/ui/grid";
-import LoadingCard from "../card/LoadingCard";
+import LoadingCard from "../card/loading-card";
 
 export default function ShimmerGrid({ size }: { size: number }) {
   return (
